+++ tests/cases/makedirs/plan.py
@@ -5,7 +5,11 @@ static("bunch.py")
 info = step("./bunch.py", out=["sub/", "sub/other/", "sub/other/foo/", "sub/other/foo/text"])
 
 # Tests for info object, only useful for testing
-if list(info.filter_out("sub/*")) != ["sub/other/"]:
+if info.filter_out("sub/*").single() != "sub/other/":
     raise AssertionError("Wrong info.filter_out")
-if list(info.filter_out("sub/other/**")) != ["sub/other/", "sub/other/foo/", "sub/other/foo/text"]:
+if info.filter_out("sub/other/**").files() != (
+    "sub/other/",
+    "sub/other/foo/",
+    "sub/other/foo/text",
+):
     raise AssertionError("Wrong info.filter_out")